Refactor data fetching & processing logic
The purpose of this branch is to re-work the way we handle data
retrieved from the server. The data that we get for a given
asset/accessory/whatever is a nested dict containing a hodge-podge of
dicts, lists, strings, ints, and nothing

Currently, the way we handle that is that we cherry-pick a hard-coded
list of expected values out of that mess and put them into a flat dict
to be used to fill out template variables in our template. It's an
inflexible solution, and requires a great deal of error handling code to
account for data that's missing or has an unexpected format.

What I'd like to do is rework this so that the whole nested list is
dynamically converted into a flat list, no matter what shape it has.